Move annotation layer renderer to own file
