+++ python/core/auto_generated/qgsmaplayer.sip.in
@@ -20,7 +20,7 @@ enum class QgsMapLayerType
   RasterLayer,
   PluginLayer,
   MeshLayer,
-  VectorTileLayer,
+  VectorTileLayer
   AnnotationLayer,
 };
