+++ src/core/annotations/qgsannotationlayerrenderer.h
@@ -1,0 +1,44 @@
+/***************************************************************************
+    qgsannotationlayerrenderer.h
+    ----------------
+    copyright            : (C) 2019 by Sandro Mani
+    email                : smani at sourcepole dot ch
+ ***************************************************************************/
+
+/***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ ***************************************************************************/
+
+#ifndef QGSANNOTATIONLAYERRENDERER_H
+#define QGSANNOTATIONLAYERRENDERER_H
+
+#define SIP_NO_FILE
+
+#include "qgis_core.h"
+#include "qgis_sip.h"
+#include "qgsmaplayerrenderer.h"
+
+class QgsAnnotationItem;
+class QgsAnnotationLayer;
+
+class CORE_EXPORT QgsAnnotationLayerRenderer : public QgsMapLayerRenderer
+{
+  public:
+
+    QgsAnnotationLayerRenderer( QgsAnnotationLayer *layer, QgsRenderContext &context );
+    ~QgsAnnotationLayerRenderer() override;
+    QgsFeedback *feedback() const override;
+    bool render() override;
+
+  private:
+    QVector< QgsAnnotationItem *> mItems;
+    std::unique_ptr< QgsFeedback > mFeedback;
+
+};
+
+#endif // QGSANNOTATIONLAYERRENDERER_H
